pyproject.toml
[project]
name = "asset-management"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mysql (>=0.0.3,<0.0.4)",
    "mysqlclient (>=2.2.6,<3.0.0)",
    "sqlalchemy (>=2.0.36,<3.0.0)",
    "mysql-connector-python (>=9.1.0,<10.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "sqlmodel (>=0.0.22,<0.0.23)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "fastapi (>=0.115.6,<0.116.0)",
    "python-jose (>=3.3.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "jwt (>=1.3.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
